# Conflicts:
#	src/main/java/com/padocadev/aplicacao/rest/pedido/PedidoControlador.java
#	src/main/java/com/padocadev/dominio/casodeuso/pedido/CriaPedidoCasoDeUso.java
#	src/main/java/com/padocadev/entities/pedido/Pedido.java
#	src/main/java/com/padocadev/infraestrutura/configuracao/BeanConfiguracao.java
#	src/test/java/com/padocadev/pedido/PedidoApiTeste.java

+++ src/main/java/com/padocadev/adapters/resposta/pedido/ItensDoPedidoResposta.java
@@ -4,7 +4,7 @@ import com.padocadev.entities.pedido.ItemPedido;
 
 import java.math.BigDecimal;
 
-public record ItensDoPedidoResposta(String nomeProduto, BigDecimal precoUnitario, int quantidade) {
+public record ItensDoPedidoResposta(String nomeProduto, BigDecimal precoUnitario, long quantidade) {
 
     public ItensDoPedidoResposta(ItemPedido itemPedido) {
         this(itemPedido.getProduto().getNome(), itemPedido.getPrecoUnitario(), itemPedido.getQuantidade());
